provide more control over scanned configurations
This commit adds two new properties to the `dependencyCheck` task:

`skipConfigurations` accepts a list of configuration names. The
dependencies of the named configurations are not analyzed. This does not
prevent the dependencies from being checked if other configurations
contain them as well - for example if other configurations extend the
named configurations.

`scanConfigurations` accepts a list of configuration names. Only the
dependencies of the named configurations will be analyzed.

`skipConfigurations` and `scanConfigurations` are mutually exclusive.

closes #6